+++ hibpcli/__init__.py
@@ -1,2 +1,2 @@
 """A command line interface for the "haveibeenpwned.com" API - speaks keepass."""
-__version__ = "0.1.0"
+__version__ = "0.2.0"